adding more style

+++ src/pages/[id].js
@@ -4,9 +4,7 @@ import { useState } from "react";
 import { useRouter } from "next/router";
 import { useSession } from "next-auth/react";
 import Layout from "../../components/layout/layout";
-import Link from "next/link";
 import CommentForm from "../../components/commentForm/commentForm";
-
 import Footer from "../../components/footer/footer";
 
 export default function RecipeDetailsPage() {
@@ -25,14 +23,18 @@ export default function RecipeDetailsPage() {
 
   console.log("userId", userId);
 
-
   const { data: allComments, mutate } = useSWR(`/api/comments?id=${id}`);
   const [comments, setComments] = useState([]);
   async function deleteRecipe() {
-    await fetch(`/api/recipes/${id}`, {
-      method: "DELETE",
-    });
-    router.push("/");
+    const confirmation = window.confirm(
+      "Are you sure you want to delete this recipe?"
+    );
+    if (confirmation) {
+      await fetch(`/api/recipes/${id}`, {
+        method: "DELETE",
+      });
+      router.push("/");
+    }
   }
   async function handleComment(e) {
     e.preventDefault();
@@ -70,21 +72,12 @@ export default function RecipeDetailsPage() {
   return (
     <>
       <Layout>
-        {userId == sessionId ? (
-          <Link href={`/edit/${data._id}`} passHref legacyBehavior>
-            <button className="border-2 p-1.5 px-6 rounded-lg">edit</button>
-          </Link>
-        ) : null}
-
-        {userId == sessionId ? (
-          <button
-            className="border-2 p-1.5 px-6 rounded-lg"
-            onClick={deleteRecipe}
-          >
-            DELETEEEEEE
-          </button>
-        ) : null}
-        <RecipeDetails data={data} />
+        <RecipeDetails
+          data={data}
+          userId={userId}
+          sessionId={sessionId}
+          handleDelete={deleteRecipe}
+        />
         <CommentForm onSubmit={handleComment} />
         <ul className="w-full  max-w-lg md:max-w-xl flex flex-col items-start space-y-2">
           {allComments &&
